Fix CodeQL warning in 64-bit term hash folding
Use an explicit 64-bit intermediate when folding bytes from uint64_t values
into the 32-bit term hash accumulator.

This avoids CodeQL's "multiplication result converted to larger type" warning
while preserving the intended uint32_t wraparound behavior of the hash.

+++ src/libAtomVM/term_hash.c
@@ -68,7 +68,7 @@ static uint32_t hash_uint32(uint32_t n, uint32_t h, uint32_t prime)
 static uint32_t hash_uint64(uint64_t n, uint32_t h, uint32_t prime)
 {
     while (n) {
-        h = h * prime + (n & 0xFF);
+        h = (uint32_t) ((uint64_t) h * prime + (n & 0xFF));
         n >>= 8;
     }
     return h * prime;
